removed size_t from generateUpdate
git-svn-id: https://btanks.svn.sourceforge.net/svnroot/btanks/trunk@836 5c0bf991-d514-0410-a65f-ce4832e571f3

+++ src/world.cpp
@@ -678,7 +678,7 @@ TRY {
 }
 
 void IWorld::generateUpdate(mrt::Serializator &s) {
-	size_t c = 0, n = _objects.size();
+	unsigned int c = 0, n = _objects.size();
 	std::set<int> skipped_objects;
 
 	for(ObjectMap::reverse_iterator i = _id2obj.rbegin(); i != _id2obj.rend(); ++i) {
